+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `impl std::error::Error for Error`.
 - Added `Copy, Clone, PartialEq, Eq` traits for `Error`.
 
+### Changed
+- Changed the return type of `iq_to_cplx` from `Complex<f32>` to `Complex<i8>`.
+
 ## [0.1.0] - 2021-03-02
 - Initial release